requirements.txt
Django==4.2.7
djangorestframework==3.14.0
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
django-cors-headers==4.3.1
Pillow==10.1.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
